Remove Grunt and dependencies from the project
We don't use them and they cause a lot of peer-dependency errors, so let's
strip Grunt and related code from the project.

A casualty of this is the `sails www` command, which tries to call Grunt to
build some stuff. Since we don't have Grunt anymore, this command can't do
anything.

As far as I can tell everything seems to still load okay; the tests pass, and
`sails lift` still works just fine.