Update Waterline to 1.1.0
The primary change here is to remove the 'alter' migration strategy from
Waterline. In addition, we remove a lot of code related to dynamic finders and
association finders, both of which we don't use.

Diff is here: https://github.com/shyp/waterline/compare/v1.0.1...v1.1.0

In addition we upgraded waterline-schema to 0.2.0, to remove 'alter' as the
default migration strategy.

https://github.com/shyp/waterline-schema/compare/shyp-master-0.1.17...v0.2.0