Remove dependency on prompt
If a project did not define a migrate strategy, Sails would prompt the user to
enter one. We don't need this setting and we don't let users choose between
strategies anymore - we only allow the 'safe' strategy.

Replaces the code that prompts the user with an error, and removes the "prompt"
dependency from package.json and npm-shrinkwrap.json.