Remove the favicon middleware
Express 4 relies on serve-favicon version 2.0.0, which requires a default
favicon; Sails doesn't have one in the repo, and it doesn't make sense for us
anyway. We serve the favicon from the filesystem, browsers rarely have a need
to navigate to api.shyp.com, and if we felt we needed a more "scalable" favicon
serving mechanism, we could easily take the dependency in the Shyp API, not
here.