Upgrade Waterline to v3.2.0
This pulls in additional (better) changes to the validation error formatting,
based on testing against the Shyp API. Also pulls in a change to clone data
before sending it to the database for INSERT/UPDATE queries.

Furthermore, there's a small change to avoid `SET`ting a primary key when
calling `.save()` on a record - only update the other fields.

We'll need to test this in the Shyp API and probably make some changes in the
API based on it.

Diff: https://github.com/Shyp/waterline/compare/v2.0.0...v3.2.0